+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.10.0
+## Added
+
+## Fixed
+
+## Changed
+* Fix error when using ipFromIpamProvider and there are no configured providers
+* Disable editor for enums with null as the only value
+
 # 1.9.0
 ## Added
 * bigip-fast-templates: Add supported version of the Microsoft ADFS template